+++ lib/utility.dart
@@ -1,0 +1,6 @@
+class Utility {
+  static String getCurrentDateAsString() {
+    DateTime now = DateTime.now();
+    return '${now.day}.${now.month}.${now.year}';
+  }
+}