+++ changelog.md
@@ -1,4 +1,9 @@
 # Changelog:
 
-## Commit 1: (version 1.0.0)
+## Commit 2: (version 1.0.1+2)
+- Added Confirmation dialog for preparing a recipe
+- Made `ConfirmationDialog` its own widget for reuseability and readability
+- Swapped order of buttons in `RecipeCreationDialog` to stay consistent (cancel button should always be the most right button)
+
+## Commit 1: (version 1.0.0+1)
 - Initial commit